refactor(inputsource): consolidate the cgo wrapper into a single file
